add IA32 etc... (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@252 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ sound/soundrom.c
@@ -67,7 +67,7 @@ void soundrom_load(UINT32 address, const char *primary) {
 	soundrom.address = address;
 }
 
-void soundrom_loadex(BYTE sw, const char *primary) {
+void soundrom_loadex(UINT sw, const char *primary) {
 
 	if (sw < 4) {
 		soundrom_load((0xc8000 + ((UINT32)sw << 14)), primary);
