+++ sound/soundrom.h
@@ -13,7 +13,7 @@ extern	SOUNDROM	soundrom;
 
 void soundrom_reset(void);
 void soundrom_load(UINT32 address, const char *primary);
-void soundrom_loadex(BYTE sw, const char *primary);
+void soundrom_loadex(UINT sw, const char *primary);
 
 #ifdef __cplusplus
 }
